don't wake up on tab close/open
+ simplify msg
