+++ src/background/navigation-manager.js
@@ -1,21 +1,29 @@
 import '/js/browser';
-import {sendTab} from '/js/msg';
+import {kAboutBlank} from '/js/consts';
 import {CHROME, FIREFOX} from '/js/ua';
 import {chromeProtectsNTP} from '/js/urls';
 import {deepEqual} from '/js/util';
 import {ignoreChromeError, MF} from '/js/util-webext';
+import {pingTab, sendTab} from './broadcast';
 import * as tabMan from './tab-manager';
 
 const listeners = new Set();
 /** @param {function(data: Object, type: ('committed'|'history'|'hash'))} fn */
 export const onUrlChange = fn => listeners.add(fn);
+/** @type {{ url: chrome.events.UrlFilter[] }} */
+export const WEB_NAV_FILTER_STYLABLE = {
+  url: [{schemes: ['http', 'https', 'file', 'ftp', 'ftps']}],
+};
 let prevData = {};
 
-chrome.webNavigation.onCommitted.addListener(onNavigation.bind('committed'));
-chrome.webNavigation.onHistoryStateUpdated.addListener(onFakeNavigation.bind('history'));
-chrome.webNavigation.onReferenceFragmentUpdated.addListener(onFakeNavigation.bind('hash'));
+chrome.webNavigation.onCommitted.addListener(onNavigation.bind(['committed']),
+  WEB_NAV_FILTER_STYLABLE);
+chrome.webNavigation.onHistoryStateUpdated.addListener(onFakeNavigation.bind(['history']),
+  WEB_NAV_FILTER_STYLABLE);
+chrome.webNavigation.onReferenceFragmentUpdated.addListener(onFakeNavigation.bind(['hash']),
+  WEB_NAV_FILTER_STYLABLE);
 
-/** @this {string} type */
+/** @this {string[]} type */
 async function onNavigation(data) {
   if (CHROME && data.timeStamp === prevData.timeStamp && deepEqual(data, prevData)) {
     return; // Chrome bug: listener is called twice with identical data
@@ -34,7 +42,7 @@ async function onNavigation(data) {
       data.url = url;
     }
   }
-  listeners.forEach(fn => fn(data, this));
+  for (const fn of listeners) fn(data, this[0]);
 }
 
 /** @this {string} type */
@@ -43,8 +51,8 @@ function onFakeNavigation(data) {
   const {tabId} = data;
   const td = tabMan.get(tabId); if (!td) return;
   const {url, frameId: f, documentId: d} = data;
-  const iid = !d && td.iid?.[f];
-  const to = d ? {documentId: d} : {frameId: f};
+  const iid = !process.env.MV3 && !d && td.iid?.[f];
+  const to = process.env.MV3 || d ? {documentId: d} : {frameId: f};
   sendTab(tabId, {method: 'urlChanged', iid, url}, to);
 }
 
@@ -85,8 +93,7 @@ if (!process.env.MV3) {
    */
   if (process.env.BUILD !== 'chrome' && FIREFOX) {
     chrome.webNavigation.onDOMContentLoaded.addListener(async ({tabId, frameId}) => {
-      if (frameId &&
-          !await sendTab(tabId, {method: 'ping'}, {frameId})) {
+      if (frameId && !await pingTab(tabId, frameId)) {
         for (const file of MF.content_scripts[0].js) {
           chrome.tabs.executeScript(tabId, {
             frameId,
@@ -96,7 +103,7 @@ if (!process.env.MV3) {
         }
       }
     }, {
-      url: [{urlEquals: 'about:blank'}],
+      url: [{urlEquals: kAboutBlank}],
     });
   }
 }
