+++ src/background/tab-util.js
@@ -1,8 +1,9 @@
 import '/js/browser';
-import {API, sendTab} from '/js/msg';
-import {CHROME, FIREFOX} from '/js/ua';
+import {API} from '/js/msg';
 import * as prefs from '/js/prefs';
+import {CHROME, FIREFOX} from '/js/ua';
 import {browserWindows, getActiveTab} from '/js/util-webext';
+import {sendTab} from './broadcast';
 
 // FF57+ supports openerTabId, but not in Android
 // (detecting FF57 by the feature it added, not navigator.ua which may be spoofed in about:config)
