+++ src/edit/on-msg-extension.js
@@ -1,3 +1,4 @@
+
 import {API, onExtension} from '/js/msg';
 import {closeCurrentTab} from '/js/util-webext';
 import editor from './editor';
