+++ src/js/consts.js
@@ -1,6 +1,14 @@
+/**
+ * WARNING!
+ * Used in limited contexts such as the offscreen document.
+ * Only for pure declarations with no side effects or marked with /*@__PURE__*/
+/** */
+
 export const UCD = 'usercssData';
+export const kAboutBlank = 'about:blank';
 export const kAppJson = 'application/json';
 export const kAppUrlencoded = 'application/x-www-form-urlencoded';
 export const kContentType = 'content-type'; // must be lowercase!
 export const kPopup = 'popup';
+export const kResolve = 'resolve';
 export const kUrl = 'url';
